editoast: remove test fixtures
